restaurant_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0

mysqlclient==2.2.0
python-decouple==3.8
Pillow==10.0.1
django-filter==23.3
drf-spectacular==0.26.5
requests==2.31.0

langchain==0.3.27
langchain-google-genai==2.1.9 
python-dotenv==1.0.0
django-cors-headers==4.7.0
